Move new entry to today's page
