+++ lib/pages/TodayPage/NewEntryPageLogic.dart
