+++ lib/widgets/Cards/EntryCard/EntryInput.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:beebetter/widgets/Cards/RecordingCard/RecordingLogic.dart';
-import 'package:beebetter/pages/NewEntryPage/NewEntryPageLogic.dart';
+import 'package:beebetter/pages/TodayPage/NewEntryPageLogic.dart';
 import 'package:beebetter/widgets/Cards/RecordingCard/RecordingCard.dart';
 import 'package:beebetter/widgets/Cards/ExpandingTextOverlay.dart';
